requirements/development.txt
Django~=3.1.12
pypiwin32~=223
pywin32~=301
Pillow~=8.3.1
requests~=2.26.0
djongo~=1.3.6
jsonpickle~=2.0.0
